[REFACTOR] RAG, youtube_service external 폴더로 분리
